move union implication instance to Implication companion object
